디렉토리 이름원복
