+++ docker/Dockerfile-bwsdb
