+++ docker/Dockerfile-bwsdbadmin
