+++ src/Aspire.AppHost/Aspire.AppHost.csproj
