+++ src/Aspire.AppHost/appsettings.json
