+++ src/Aspire.ServiceDefaults/Aspire.ServiceDefaults.csproj
