+++ src/Python.FastApi/common/system/constants.py
