+++ src/Python.FastApi/engage/algebra/algebra.py
