+++ src/Python.FastApi/resources/config.properties
